# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mountain Gear Rental API",
        "description": "API documentation for Mountain Gear Rental - Hiking Equipment Rental System. This API provides endpoints for managing hiking equipment rentals, customer data, and rental transactions.",
        "termsOfService": "https://pemweb.uas/terms",
        "contact": {
            "name": "Mountain Gear Support",
            "email": "info@mountaingear.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://pemweb.uas",
            "description": "Mountain Gear Rental API Server"
        }
    ],
    "paths": {
...
        "/api/rentals": {
            "get": {
                "tags": [
                    "Rentals"
                ],
                "summary": "Get all rentals",
                "description": "Returns a list of all rental transactions",
                "operationId": "getRentals",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by rental status",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pending",
                                "confirmed",
                                "active",
                                "completed",
                                "cancelled"
                            ]
                        }
                    },
                    {
                        "name": "customer_id",
                        "in": "query",
                        "description": "Filter by customer ID",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Rentals retrieved successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Rental"
                                            }
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 50
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Rentals"
                ],
                "summary": "Create new rental",
                "description": "Create a new equipment rental booking",
                "operationId": "createRental",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "customer_id",
                                    "start_date",
                                    "end_date",
                                    "equipment_items"
                                ],
                                "properties": {
                                    "customer_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-07-25"
                                    },
                                    "end_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-07-30"
                                    },
                                    "notes": {
                                        "type": "string",
                                        "example": "Weekend hiking trip to Mount Bromo"
                                    },
                                    "equipment_items": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "hiking_equipment_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "quantity": {
                                                    "type": "integer",
                                                    "example": 2
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Rental created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Rental created successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Rental"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
        "/api/rentals/{id}": {
            "get": {
                "tags": [
                    "Rentals"
                ],
                "summary": "Get rental by ID",
                "description": "Returns details of specific rental",
                "operationId": "getRentalById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Rental ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Rental retrieved successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Rental"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Rental not found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
        "/api/rentals/{id}/status": {
            "put": {
                "tags": [
                    "Rentals"
                ],
                "summary": "Update rental status",
                "description": "Update the status of a rental (confirm, start, complete, cancel)",
                "operationId": "updateRentalStatus",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Rental ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "confirmed",
                                            "active",
                                            "completed",
                                            "cancelled"
                                        ],
                                        "example": "confirmed"
                                    },
                                    "notes": {
                                        "type": "string",
                                        "example": "Payment confirmed, equipment ready for pickup"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Rental status updated successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Rental"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Rental not found"
                    },
                    "422": {
                        "description": "Invalid status transition"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Rental": {
                "title": "Equipment Rental",
                "description": "Rental transaction model",
                "required": [
                    "id",
                    "customer_id",
                    "rental_code",
                    "start_date",
                    "end_date",
                    "total_amount",
                    "status"
                ],
                "properties": {
                    "id": {
                        "description": "Rental ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "customer_id": {
                        "description": "Customer ID",
                        "type": "integer",
                        "example": 1
                    },
                    "customer": {
                        "$ref": "#/components/schemas/Customer"
                    },
                    "rental_code": {
                        "description": "Unique rental code",
                        "type": "string",
                        "example": "RNT-ABC123"
                    },
                    "start_date": {
                        "description": "Rental start date",
                        "type": "string",
                        "format": "date",
                        "example": "2025-07-25"
                    },
                    "end_date": {
                        "description": "Rental end date",
                        "type": "string",
                        "format": "date",
                        "example": "2025-07-30"
                    },
                    "total_amount": {
                        "description": "Total rental amount in IDR",
                        "type": "number",
                        "format": "float",
                        "example": 750000
                    },
                    "status": {
                        "description": "Rental status",
                        "type": "string",
                        "enum": [
                            "pending",
                            "confirmed",
                            "active",
                            "completed",
                            "cancelled"
                        ],
                        "example": "confirmed"
                    },
                    "notes": {
                        "description": "Additional notes for the rental",
                        "type": "string",
                        "example": "Weekend hiking trip to Mount Bromo"
                    },
                    "rental_items": {
                        "description": "List of rented equipment items",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/RentalItem"
                        }
                    },
                    "payments": {
                        "description": "Payment records for this rental",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Payment"
                        }
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-07-21T10:00:00.000000Z"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-07-21T10:00:00.000000Z"
                    }
                },
                "type": "object"
            },
            "RentalItem": {
                "title": "Rental Item",
                "description": "Individual item in a rental",
                "required": [
                    "id",
                    "rental_id",
                    "hiking_equipment_id",
                    "quantity",
                    "price_per_day",
                    "subtotal"
                ],
                "properties": {
                    "id": {
                        "description": "Rental item ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "rental_id": {
                        "description": "Rental ID",
                        "type": "integer",
                        "example": 1
                    },
                    "hiking_equipment_id": {
                        "description": "Equipment ID",
                        "type": "integer",
                        "example": 1
                    },
                    "hiking_equipment": {
                        "$ref": "#/components/schemas/HikingEquipment"
                    },
                    "quantity": {
                        "description": "Quantity rented",
                        "type": "integer",
                        "example": 2
                    },
                    "price_per_day": {
                        "description": "Price per day at time of rental",
                        "type": "number",
                        "format": "float",
                        "example": 25000
                    },
                    "subtotal": {
                        "description": "Subtotal for this item",
                        "type": "number",
                        "format": "float",
                        "example": 300000
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-07-21T10:00:00.000000Z"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-07-21T10:00:00.000000Z"
                    }
                },
                "type": "object"
            },
            "Customer": {
                "title": "Customer",
                "description": "Customer model",
                "required": [
                    "id",
                    "name",
                    "email",
                    "phone"
                ],
                "properties": {
                    "id": {
                        "description": "Customer ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "description": "Customer full name",
                        "type": "string",
                        "maxLength": 255,
                        "example": "John Doe"
                    },
                    "email": {
                        "description": "Customer email address",
                        "type": "string",
                        "format": "email",
                        "example": "john.doe@example.com"
                    },
                    "phone": {
                        "description": "Customer phone number",
                        "type": "string",
                        "example": "+62812345678"
                    },
                    "address": {
                        "description": "Customer address",
                        "type": "string",
                        "example": "Jl. Sudirman No. 123, Jakarta"
                    },
                    "id_card_number": {
                        "description": "Customer ID card number",
                        "type": "string",
                        "example": "3201234567890123"
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-07-21T10:00:00.000000Z"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-07-21T10:00:00.000000Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "ApiKeyAuth": {
                "type": "apiKey",
                "description": "API Key Authentication",
                "name": "X-API-KEY",
                "in": "header"
            }
        }
    },
...
}